% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setting.R
\name{get_setting_grade_max}
\alias{get_setting_grade_max}
\title{Get Maximum Grade of Quiz Settting}
\usage{
get_setting_grade_max(df_raw)
}
\arguments{
\item{df_raw}{A data.frame of Moodle Quiz report}
}
\value{
Numeric vector of length 1 indicate maximum quiz grade
}
\description{
Get a maximum grade setting of any Moodle quiz report.
}
