% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_setting.R
\name{get_setting_questions_no_max}
\alias{get_setting_questions_no_max}
\title{Get Question Number & Maximum Score}
\usage{
get_setting_questions_no_max(df_gr)
}
\arguments{
\item{df_gr}{A data.frame of Moodle Grades report}
}
\value{
A numeric vector represents maximum score of each questions (sorted from low to high).
}
\description{
For Moodle Grades file, get a question's maximum score corresponding to question's number (sorted from the first to the last question).
}
