% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grades_df}
\alias{grades_df}
\title{Moodle Grades Report}
\format{
A data frame with 461 rows and 19 variables:
\describe{
\item{Surname}{Student Surname (simulated)}
\item{First name}{Student Firstname (simulated)}
\item{Institution}{Student Institution (leaves as \code{NA})}
\item{Department}{Student Department (leaves as \code{NA})}
\item{Email address}{Student Email (simulated)}
\item{State}{State of the quiz, one of "Finished" or "In progress"}
\item{Started on}{Quiz started time}
\item{Completed}{Quiz submitted time}
\item{Time taken}{Time duration of doing quiz}
\item{Grade/9.00}{Quiz Score of each student (max setting is 9.00)}
\item{Q. 1 /1.00}{Score of Question 1}
\item{Q. 2 /1.00}{Score of Question 2}
\item{Q. 3 /1.00}{Score of Question 3}
\item{Q. 4 /1.00}{Score of Question 4}
\item{Q. 5 /1.00}{Score of Question 5}
\item{Q. 6 /1.00}{Score of Question 6}
\item{Q. 7 /1.00}{Score of Question 7}
\item{Q. 8 /1.00}{Score of Question 8}
\item{Q. 9 /1.00}{Score of Question 9}
}
}
\source{
\url{https://docs.moodle.org/311/en/Quiz_reports}
}
\usage{
grades_df
}
\description{
A simulated data frame of \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Grades Report}
}
\keyword{datasets}
