% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{item_discrim}
\alias{item_discrim}
\title{Item Discrimination of Moodle Grade Report}
\usage{
item_discrim(
  data,
  cor_method = c("pearson", "kendall", "spearman"),
  conf.level = 0.95
)
}
\arguments{
\item{data}{(GradesReport) A data frame of class "GradesReport"}

\item{cor_method}{(Character) A character string indicating which correlation coefficient is to be used for the test. One of "pearson", "kendall", or "spearman", can be abbreviated.}

\item{conf.level}{(Numeric) Confidence level for the returned confidence interval. Currently only used for the Pearson product moment correlation coefficient if there are at least 4 complete pairs of observations.}
}
\value{
A data frame of Item Discrimination Index for each questions.
}
\description{
Calculate \href{https://www.proftesting.com/test_topics/steps_9.php}{Item Discrimination Index} of \href{https://docs.moodle.org/311/en/Quiz_reports#Grades_report}{Moodle Grade Report}
(see \strong{Details}).
}
\details{
\strong{Item discrimination index (r)} is a measure of how well an item is able to distinguish between examinees who are knowledgeable and those who are not.
It is a pairwise \strong{point-biserial correlation} between the \emph{score of each questions} ("Q" columns) and \emph{total score of the quiz} ("Grade" column).

\code{\link[=item_discrim]{item_discrim()}} calculate the correlation by \code{\link[stats:cor.test]{stats::cor.test()}} function using Pearson's correlation coefficient (by default),
then summarize each parameters into a data frame by \code{\link[broom:reexports]{broom::tidy()}}. Different types of correlation can be specified by \code{cor_method} argument.
}
\examples{
library(moodleStats)
# Prepare
grades_df_preped <- prep_grades_report(grades_df)

# Calculate
item_discrim(grades_df_preped)
}
\references{
\itemize{
\item What is \href{https://www.proftesting.com/test_topics/steps_9.php}{Item Discrimination Index}.
}
}
