% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleStats-package.R
\docType{package}
\name{moodleStats-package}
\alias{moodleStats}
\alias{moodleStats-package}
\title{moodleStats: Analysis of Moodle Quiz Report}
\description{
This package contain funtions that perform analyis of moodle quiz report to get quiz statistics and quiz item analysis.
}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{ki11ip0.s.a.s@gmail.com}

}
\keyword{internal}
