% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{moodleStats_example}
\alias{moodleStats_example}
\title{Get path to moodleStats example}
\usage{
moodleStats_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
moodleStats comes bundled with a number of sample files in its \code{inst/extdata}
directory. This function make them easy to access
}
\examples{
library(moodleStats)
moodleStats_example()
moodleStats_example("grades_report.csv")
}
