% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep_grades_report}
\alias{prep_grades_report}
\title{Prepare Moodle Grades Report for Analysis}
\usage{
prep_grades_report(
  data,
  sep_name = " ",
  choose_state = c("Finished", "In progress", "all"),
  choose_grade = c("max", "min", "all"),
  choose_time = c("first", "last", "all")
)
}
\arguments{
\item{data}{(Data Frame) A data frame of \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Grades Report}}

\item{sep_name}{(Character) Separator of "First name" and "Surname" column.}

\item{choose_state}{(Character) Options to filter a "State" column, must be one of:
\itemize{
\item \strong{Finished:} (default) choose only the "Finished" attempts.
\item \strong{In progress:} choose only the "In progress" attempts.
\item \strong{all:} no filter applied, choose all attempts.
}}

\item{choose_grade}{(Character) Options to filter a "Grade/x.xx" column, must be one of:
\itemize{
\item \strong{max:} (default) choose attempts that has maximum score of each students
\item \strong{min:} choose attempts that has minimum score of each students
\item \strong{all:} no filter applied, choose all attempts.
}}

\item{choose_time}{(Character) Options to filter a "Started on" or "Started" column, must be one of:
\itemize{
\item \strong{first:} (default) choose first attempt of each students
\item \strong{last:} choose last attempt of each students
\item \strong{all:} no filter applied, choose all attempts.
}}
}
\value{
A cleaned and (optionally) filtered data frame with class "GradesReport" and "MoodleQuizReport".
}
\description{
Prepare a \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Grades Report} data frame
for further analysis (see \strong{Details}).
}
\details{
Steps to prepare a \href{https://docs.moodle.org/311/en/Quiz_reports}{Moodle Grades Report} data frame:
\itemize{
\item Clean some column names
\item Join "First name" and "Surname" by \code{sep_name}
\item Format "Grade/x.xx" column to numeric ("Not yet graded" and "-" will be \code{NA})
\item Format "Q. x /x.xx" column to numeric ("Requires grading" and "-" will be \code{NA})
\item Filter a "State" column by \code{choose_state}
\item Filter a "Grade" column by \code{choose_grade} (after \code{choose_state} has been applied)
\item Filter a "Started on" or "Started" column by \code{choose_time} (after \code{choose_grade} has been applied)
}
}
\examples{
prep_grades_report(grades_df)
}
