% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{questions_stats}
\alias{questions_stats}
\title{Question Stats of Moodle Grade Report}
\usage{
questions_stats(data, show = NULL)
}
\arguments{
\item{data}{(GradesReport) A data frame of class "GradesReport"}

\item{show}{If \code{NULL}, show every parameters.
Otherwise, specify which summary statistics you want to show as a character vector, any of:
"n", "min", "max", "median", "Q1", "Q3", "IQR", "MAD", "mean", "SD", "skewness", and "kurtosis" (for \emph{excess} kurtosis).}
}
\value{
a data frame of summary statistics
}
\description{
Calculate summary statistics for each questions of \href{https://docs.moodle.org/311/en/Quiz_reports#Grades_report}{Moodle Grade Report} (See \strong{Details}).
}
\details{
\code{\link[=questions_stats]{questions_stats()}} calculate grouped summary statistics by questions ("Q").
You can specify which parameters to calculate by \code{show} argument.
}
\examples{
library(moodleStats)
# Prepare
grades_df_preped <- prep_grades_report(grades_df)

# Calculate
questions_stats(grades_df_preped, show = c("n", "min", "max", "mean"))
}
