% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz_meta.R
\name{quiz_meta}
\alias{quiz_meta}
\title{Get Moodle Quiz Metadata}
\usage{
quiz_meta(data)
}
\arguments{
\item{data}{a Moodle Quiz Report data frame}
}
\value{
Information printed to R console, and a list that has elements corresponding to each information (invisibly).
}
\description{
Display Moodle quiz report metadata to R console.
}
\details{
This function displays
\itemize{
\item \strong{Moodle Quiz Setting:} a quiz setting in Moodle such as maximum grades possible of the quiz
and maximum score possible of each question (if it is a Moodle Grades Report).
\item \strong{Data Frame Filtering:} If Moodle Quiz was passed to \code{prep_} function family,
this will show what filters have been applied to the data frame.
}
}
\examples{
# Quiz Settings
quiz_meta(grades_df)
# Quiz Settings with Filtering
grades_df_preped <- prep_grades_report(grades_df)
quiz_meta(grades_df_preped)
}
