% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{summary_questions}
\alias{summary_questions}
\title{Questions Summary of Moodle Grade Report}
\usage{
summary_questions(data, cor_method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{data}{(GradesReport) A data frame of class "GradesReport"}

\item{cor_method}{(Character) A character string indicating which correlation coefficient is to be used for calculating \code{Discrimination_Index}.
Default is Pearson's correlation ("pearson"). Other types can be specified such as "kendall", or "spearman", can be abbreviated.}
}
\value{
a data frame of overall summary per questions
}
\description{
Calculate questions summary and item analysis of \href{https://docs.moodle.org/311/en/Quiz_reports#Grades_report}{Moodle Grade Report} (See \strong{Details}).
}
\details{
\code{\link[=summary_questions]{summary_questions()}} calculate overall summary grouped by each questions of Moodle Grade Report, and return the result as a data frame.
The data frame has a column "Questions" for questions number in the quiz. Other columns are the followings.

\strong{Basic summary statistic}
\itemize{
\item \strong{n}: number of students who answered each questions.
\item \strong{min}: student's minimum score of each questions.
\item \strong{max}: student's maximum score of each questions.
\item \strong{max_setting}: maximum score possible of each questions (from the quiz setting).
\item \strong{mean}: mean score of each questions.
\item \strong{SD}: standard deviation of each questions score.
}

\strong{Item analysis}
\itemize{
\item \strong{Difficulty_Index}: Item Difficulty Index (p) is a measure of the \emph{proportion} of examinees who answered the item correctly.
It ranges between 0.0 and 1.0, \emph{higher} value indicate \emph{lower} question difficulty, and vice versa.
\item \strong{Discrimination_Index}: Item Discrimination Index (r) is a measure of how well an item is able to distinguish between examinees who are knowledgeable and those who are not.
It is a pairwise \strong{point-biserial correlation} between the \emph{score of each questions} ("Q" columns) and \emph{total score of the quiz} ("Grade" column).
It range between -1.0 to 1.0. Negative values suggest a problem, indicating that score of the particular question is negatively correlated with total quiz score; therefore, revision of the question is suggested.
\item \strong{p.value}: A level of significant (p-value) of \code{Discrimination_Index}.
\item \strong{p.signif}: A Symbol indicating level of significant of \code{Discrimination_Index}.
}
}
\examples{
# Prepare
grades_df_preped <- prep_grades_report(grades_df)

# Question Summary
summary_questions(grades_df_preped)
}
\references{
\itemize{
\item \href{https://www.proftesting.com/test_topics/steps_9.php}{How to interpret an item analysis}
}
}
\seealso{
\itemize{
\item \code{\link[=summary_quiz]{summary_quiz()}} for quiz summary statistics
\item \code{\link[=questions_stats]{questions_stats()}} for question summary statistics with more parameters.
\item \code{\link[=item_discrim]{item_discrim()}} for Item discrimination index with more parameters.
}
}
