% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiz.R
\name{summary_quiz}
\alias{summary_quiz}
\title{Quiz Summary of Moodle Grade Report}
\usage{
summary_quiz(data)
}
\arguments{
\item{data}{(GradesReport) A data frame of class "GradesReport" (May expand this to other class in the future)}
}
\value{
a one-row data frame of quiz summary statistics
}
\description{
Calculate quiz summary statistics of \href{https://docs.moodle.org/311/en/Quiz_reports#Grades_report}{Moodle Grade Report} (See \strong{Details}).
}
\details{
\code{\link[=summary_quiz]{summary_quiz()}} is a generic function that calculate quiz summary statistics
of "GradesReport" data frame.

The result is one row data frame with the following columns:
\itemize{
\item \strong{n}: number of students in the quiz
\item \strong{min}: student's minimum score
\item \strong{max}: student's maximum score
\item \strong{max_quiz}: maximum score possible (from the quiz setting)
\item \strong{median}: student's median score
\item \strong{Q1}: 1st quartile of student's score (see \code{\link[stats:quantile]{stats::quantile()}})
\item \strong{Q3}: 3rd quartile of student's score (see \code{\link[stats:quantile]{stats::quantile()}})
\item \strong{IQR}: Interquartile range of student's score (see \code{\link[stats:IQR]{stats::IQR()}})
\item \strong{MAD}: Median Absolute Deviation of student's score (see \code{\link[stats:mad]{stats::mad()}})
\item \strong{mean}: student's mean score
\item \strong{SD}: standard deviation of student's score
\item \strong{skewness}: a measure of asymmetry, \emph{positive} values indicate that the tail is on the \emph{right}, whereas \emph{negative} values indicate that the tail is on the \emph{left}  (see \code{\link[moments:skewness]{moments::skewness()}}).
\item \strong{kurtosis}: \emph{excess kurtosis}, \emph{positive} values (leptokurtic) indicate that score distribution has a \emph{fatter} tails (more outliers), whereas \emph{negative} values (platykurtic) indicate that score distribution has a \emph{thinner} tails (less outliers).
Please note that \verb{excess kurtosis = kurtosis - 3}. (see \code{\link[moments:skewness]{moments::skewness()}} for Pearson's measure of kurtosis).
\item \strong{Cronbach_Alpha}: a measure of internal consistency reliability of quiz. It is an average correlation of all question's score ("Q" column). Higher value indicates good overall correlation, and vice versa. (see \code{\link[DescTools:CronbachAlpha]{DescTools::CronbachAlpha()}})
}
}
\examples{
# Prepare Quiz
grades_df_preped <- prep_grades_report(grades_df)

# Quiz Summary Statistics
summary_quiz(grades_df_preped)
}
\seealso{
\itemize{
\item \code{\link[=summary_questions]{summary_questions()}} for question's summary statistics.
}
}
