% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleStats-package.R
\docType{package}
\name{moodleStats-package}
\alias{moodleStats}
\alias{moodleStats-package}
\title{moodleStats: Analysis of Moodle Quiz Report}
\description{
This package contain funtions that perform analyis of moodle quiz report to get quiz statistics and quiz item analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Lightbridge-KS/moodleStats}
  \item \url{https://lightbridge-ks.github.io/moodleStats/}
  \item Report bugs at \url{https://github.com/Lightbridge-KS/moodleStats/issues}
}

}
\author{
\strong{Maintainer}: Kittipos Sirivongrungson \email{ki11ip0.s.a.s@gmail.com}

}
\keyword{internal}
